refactor: enums folder
